refactor utils and imports
